last ditch attempt to deflake watchdog
