+++ hypothesis-python/tests/watchdog/test_database.py
@@ -8,12 +8,9 @@
 # v. 2.0. If a copy of the MPL was not distributed with this file, You can
 # obtain one at https://mozilla.org/MPL/2.0/.
 
-import math
 import sys
 import time
 from collections import Counter
-
-import pytest
 
 from hypothesis import Phase, settings
 from hypothesis.database import (
@@ -21,9 +18,8 @@ from hypothesis.database import (
     InMemoryExampleDatabase,
     MultiplexedDatabase,
 )
-from hypothesis.internal.reflection import get_pretty_function_description
 
-from tests.common.utils import flaky, skipif_threading
+from tests.common.utils import skipif_threading, wait_for
 from tests.cover.test_database_backend import _database_conforms_to_listener_api
 
 # we need real time here, not monkeypatched for CI
@@ -31,15 +27,16 @@ time_sleep = time.sleep
 
 
 def test_database_listener_directory():
-    # this test is very expensive because we wait between every rule for the
-    # filesystem observer to fire. Limit examples/step count as much as possible.
     _database_conforms_to_listener_api(
         lambda path: DirectoryBasedExampleDatabase(path),
-        flush=lambda _db: time_sleep(0.2),
         supports_value_delete=False,
-        # expensive flush makes shrinking take forever
         parent_settings=settings(
-            max_examples=5, stateful_step_count=10, phases=set(Phase) - {Phase.shrink}
+            # this test is very expensive because we wait between every rule for
+            # the filesystem observer to fire.
+            max_examples=5,
+            stateful_step_count=10,
+            # expensive runtime makes shrinking take forever
+            phases=set(Phase) - {Phase.shrink},
         ),
     )
 
@@ -47,7 +44,6 @@ def test_database_listener_directory():
 # seen flaky on test-win; we get *three* of the same save events in the first
 # assertion, which...is baffling, and possibly a genuine bug (most likely in
 # watchdog).
-@flaky(max_runs=5, min_passes=1)
 @skipif_threading  # add_listener is not thread safe because watchdog is not
 def test_database_listener_multiplexed(tmp_path):
     db = MultiplexedDatabase(
@@ -61,45 +57,33 @@ def test_database_listener_multiplexed(tmp_path):
     db.add_listener(listener)
 
     db.save(b"a", b"a")
-    time_sleep(0.2)
-    assert events == [("save", (b"a", b"a"))] * 2
+    wait_for(lambda: events == [("save", (b"a", b"a"))] * 2, timeout=60)
 
     db.remove_listener(listener)
     db.delete(b"a", b"a")
     db.save(b"a", b"b")
-    time_sleep(0.2)
-    assert events == [("save", (b"a", b"a"))] * 2
+    wait_for(lambda: events == [("save", (b"a", b"a"))] * 2, timeout=60)
 
     db.add_listener(listener)
     db.delete(b"a", b"b")
     db.save(b"a", b"c")
-    time_sleep(0.2)
     # InMemory database fires immediately, while DirectoryBased has to
     # wait for filesystem listeners. Therefore the events can arrive out of
     # order. Test a weaker multiset property, disregarding ordering.
-    assert Counter(events[2:]) == {
-        # InMemory
-        ("delete", (b"a", b"b")): 1,
-        # DirectoryBased
-        ("delete", (b"a", None)): 1,
-        # both
-        ("save", (b"a", b"c")): 2,
-    }
-
-
-def wait_for(condition, *, timeout=1, interval=0.01):
-    for _ in range(math.ceil(timeout / interval)):
-        if condition():
-            return
-        time_sleep(interval)
-    raise Exception(
-        f"timing out after waiting {timeout}s for condition "
-        f"{get_pretty_function_description(condition)}"
+    wait_for(
+        lambda: Counter(events[2:])
+        == {
+            # InMemory
+            ("delete", (b"a", b"b")): 1,
+            # DirectoryBased
+            ("delete", (b"a", None)): 1,
+            # both
+            ("save", (b"a", b"c")): 2,
+        },
+        timeout=60,
     )
 
 
-# seen flaky on check-coverage (timeout in first wait_for)
-@flaky(max_runs=5, min_passes=1)
 @skipif_threading  # add_listener is not thread safe because watchdog is not
 def test_database_listener_directory_explicit(tmp_path):
     db = DirectoryBasedExampleDatabase(tmp_path)
@@ -111,13 +95,12 @@ def test_database_listener_directory_explicit(tmp_path):
     db.add_listener(listener)
 
     db.save(b"k1", b"v1")
-    wait_for(lambda: events == [("save", (b"k1", b"v1"))], timeout=5)
+    wait_for(lambda: events == [("save", (b"k1", b"v1"))], timeout=60)
 
     db.remove_listener(listener)
     db.delete(b"k1", b"v1")
     db.save(b"k1", b"v2")
-    time_sleep(0.2)
-    assert events == [("save", (b"k1", b"v1"))]
+    wait_for(lambda: events == [("save", (b"k1", b"v1"))], timeout=60)
 
     db.add_listener(listener)
     db.delete(b"k1", b"v2")
@@ -128,7 +111,7 @@ def test_database_listener_directory_explicit(tmp_path):
             ("delete", (b"k1", None)),
             ("save", (b"k1", b"v3")),
         ],
-        timeout=5,
+        timeout=60,
     )
 
     # moving into a nonexistent key
@@ -171,11 +154,6 @@ def test_database_listener_directory_explicit(tmp_path):
         raise NotImplementedError(f"unknown platform {sys.platform}")
 
 
-# seen flaky on windows CI (timeout in wait_for).
-# when this happens it seems to occur consistently within that run, so the
-# @flaky doesn't help.
-@pytest.mark.skipif(sys.platform.startswith("win"), reason="too flaky on windows")
-@flaky(max_runs=5, min_passes=1)
 @skipif_threading  # add_listener is not thread safe because watchdog is not
 def test_database_listener_directory_move(tmp_path):
     db = DirectoryBasedExampleDatabase(tmp_path)
@@ -201,7 +179,7 @@ def test_database_listener_directory_move(tmp_path):
             # windows doesn't fire move events, so value is None
             ("delete", (b"k1", None if sys.platform.startswith("win") else b"v1")),
         },
-        timeout=5,
+        timeout=60,
     )
 
 
@@ -223,5 +201,4 @@ def test_still_listens_if_directory_did_not_exist(tmp_path):
 
     assert not events
     db.save(b"k1", b"v1")
-    time_sleep(0.2)
-    assert len(events) == 1
+    wait_for(lambda: len(events) == 1, timeout=60)
